Worked on player management code and player mover.
Players are now registered with the GameManager when joining, eleminating the need for the PlayerManager class. PlayerObjects are created when a new game manager is created allowing for variants of player objects in different scenes.

The player mover now deals with player objects and should technically be working at a base level. Needs to be tested.